Remove bootstrap app.
This functionality was folded into the platform shortly before release.
Unfortunately, when I synced the google3 code to staging before release,
I just copied the (with `cp -r`) the files from google3 to the staging
repository, meaning the files that had been deleted in google3 were not
removed in the staging repository, so they made it into the open souce
repository. I only noticed this when trying to sync from github to
google3 using copybara.